feat(nvs): refactor espp::Nvs to use espp::NvsHandle
