+++ components/nvs/include/nvs_handle_espp.hpp
@@ -2,12 +2,14 @@
 
 #include <string>
 
-#include "nvs.h"
-#include "nvs.hpp"
-#include "nvs_flash.h"
-#include "nvs_handle.hpp"
+#include <nvs.h>
+#include <nvs.hpp>
+#include <nvs_flash.h>
+#include <nvs_handle.hpp>
 
 #include "base_component.hpp"
+#include "nvs_errc.hpp"
+#include "nvs_handle_espp.hpp"
 
 namespace espp {
 /**
@@ -17,17 +19,17 @@ namespace espp {
  * encapsulates all direct interactions with the NVS to ensure proper error handling
  * and namespace management.
  *
- * @section nvshandle_ex1 NVSHandle Example
+ * @section nvshandle_ex1 NvsHandle Example
  * @snippet nvs_example.cpp nvshandle example
  */
-class NVSHandle : public BaseComponent {
+class NvsHandle : public BaseComponent {
 public:
-  /// @brief Construct a new NVSHandle object
+  /// @brief Construct a new NvsHandle object
   /// @param[in] ns_name Namespace for NVS
   /// @param[out] ec Saves a std::error_code representing success or failure
-  /// @details Create an NVSHandle object for the key-value pairs in the ns_name namespace
-  explicit NVSHandle(const char *ns_name, std::error_code &ec)
-      : BaseComponent("NVSHandle", espp::Logger::Verbosity::WARN) {
+  /// @details Create an NvsHandle object for the key-value pairs in the ns_name namespace
+  explicit NvsHandle(const char *ns_name, std::error_code &ec)
+      : BaseComponent("NvsHandle", espp::Logger::Verbosity::WARN) {
     if (strlen(ns_name) > 15) {
       logger_.error("Namespace too long, must be <= 15 characters: {}", ns_name);
       ec = make_error_code(NvsErrc::Namespace_Length_Too_Long);
@@ -35,7 +37,7 @@ public:
     }
 
     esp_err_t err;
-    handle = nvs::open_nvs_handle(ns_name, NVS_READWRITE, &err);
+    handle_ = nvs::open_nvs_handle(ns_name, NVS_READWRITE, &err);
     if (err != ESP_OK) {
       logger_.error("Error {} opening NVS handle for namespace '{}'!", esp_err_to_name(err),
                     ns_name);
@@ -45,20 +47,19 @@ public:
 
   /// @brief Reads a variable from the NVS
   /// @param[in] key NVS Key of the variable to read
-  /// @param[in] value Variable to read
+  /// @param[out] value Variable to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   /// @details Reads the value of key into value, if key exists
   template <typename T> void get(const char *key, T &value, std::error_code &ec) {
-    check_key_length(key, ec);
-    if (ec)
+    if (!check_handle_initialized(ec))
       return;
 
-    if (!check_handle_initialized(ec))
+    if (!check_key(key, ec))
       return;
 
     esp_err_t err;
     T readvalue;
-    err = handle->get_item(key, readvalue);
+    err = handle_->get_item(key, readvalue);
     switch (err) {
     case ESP_OK:
       value = readvalue;
@@ -76,7 +77,7 @@ public:
 
   /// @brief Reads a variable from the NVS
   /// @param[in] key NVS Key of the variable to read
-  /// @param[in] value Variable to read
+  /// @param[out] value Variable to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   template <typename T> void get(std::string_view key, T &value, std::error_code &ec) {
     get(key.data(), value, ec);
@@ -84,7 +85,7 @@ public:
 
   /// @brief Reads a bool from the NVS
   /// @param[in] key NVS Key of the bool to read
-  /// @param[in] value bool to read
+  /// @param[out] value bool to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   /// @details Read the key/variable pair
   void get(const char *key, bool &value, std::error_code &ec) {
@@ -96,13 +97,13 @@ public:
 
   /// @brief Reads a bool from the NVS
   /// @param[in] key NVS Key of the bool to read
-  /// @param[in] value bool to read
+  /// @param[out] value bool to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   void get(std::string_view key, bool &value, std::error_code &ec) { get(key.data(), value, ec); }
 
   /// @brief Reads a string from the NVS
   /// @param[in] key NVS Key of the string to read
-  /// @param[in] value string to read
+  /// @param[inout] value string to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   void get(std::string_view key, std::string &value, std::error_code &ec) {
     get(key.data(), value, ec);
@@ -110,26 +111,25 @@ public:
 
   /// @brief Reads a string from the NVS
   /// @param[in] key NVS Key of the string to read
-  /// @param[in] value string to read
+  /// @param[out] value string to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   void get(const char *key, std::string &value, std::error_code &ec) {
-    check_key_length(key, ec);
-    if (ec)
+    if (!check_handle_initialized(ec))
       return;
 
-    if (!check_handle_initialized(ec))
+    if (!check_key(key, ec))
       return;
 
     esp_err_t err;
     std::size_t len = 0;
-    err = handle->get_item_size(nvs::ItemType::SZ, key, len);
+    err = handle_->get_item_size(nvs::ItemType::SZ, key, len);
     if (err != ESP_OK) {
       logger_.error("Error {} reading!", esp_err_to_name(err));
       ec = make_error_code(NvsErrc::Read_NVS_Failed);
       return;
     }
     value.resize(len);
-    err = handle->get_string(key, value.data(), len);
+    err = handle_->get_string(key, value.data(), len);
     if (err != ESP_OK) {
       ec = make_error_code(NvsErrc::Read_NVS_Failed);
       logger_.error("Error {} reading from NVS!", esp_err_to_name(err));
@@ -138,21 +138,131 @@ public:
     return;
   }
 
+  /// @brief Reads a variable from the NVS
+  /// @param[in] key NVS Key of the variable to read
+  /// @param[out] value Variable to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  /// @details Reads the value of key into value, if key exists
+  template <typename T>
+  void get(const char *key, T &value, const T default_value, std::error_code &ec) {
+    if (!check_handle_initialized(ec))
+      return;
+
+    if (!check_key(key, ec))
+      return;
+
+    esp_err_t err;
+    T readvalue;
+    err = handle_->get_item(key, readvalue);
+    switch (err) {
+    case ESP_OK:
+      value = readvalue;
+      break;
+    case ESP_ERR_NVS_NOT_FOUND:
+      logger_.warn("The value is not initialized yet! key '{}' set to: {}", key, default_value);
+      set(key, default_value, ec);
+      if (ec)
+        return;
+      value = default_value;
+      break;
+    default:
+      logger_.error("Error {} reading!", esp_err_to_name(err));
+      ec = make_error_code(NvsErrc::Read_NVS_Failed);
+    }
+    return;
+  }
+
+  /// @brief Reads a variable from the NVS
+  /// @param[in] key NVS Key of the variable to read
+  /// @param[out] value Variable to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  template <typename T>
+  void get(std::string_view key, T &value, const T default_value, std::error_code &ec) {
+    get(key.data(), value, default_value, ec);
+  }
+
+  /// @brief Reads a bool from the NVS
+  /// @param[in] key NVS Key of the bool to read
+  /// @param[out] value bool to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  /// @details Read the key/variable pair
+  void get(const char *key, bool &value, const bool default_value, std::error_code &ec) {
+    uint8_t u8 = static_cast<uint8_t>(value);
+    uint8_t u8_default = static_cast<uint8_t>(default_value);
+    get<uint8_t>(key, u8, u8_default, ec);
+    if (!ec)
+      value = static_cast<bool>(u8);
+  }
+
+  /// @brief Reads a bool from the NVS
+  /// @param[in] key NVS Key of the bool to read
+  /// @param[out] value bool to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  void get(std::string_view key, bool &value, const bool default_value, std::error_code &ec) {
+    get(key.data(), value, default_value, ec);
+  }
+
+  /// @brief Reads a string from the NVS
+  /// @param[in] key NVS Key of the string to read
+  /// @param[out] value string to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  void get(std::string_view key, std::string &value, const std::string &default_value,
+           std::error_code &ec) {
+    get(key.data(), value, default_value, ec);
+  }
+
+  /// @brief Reads a string from the NVS
+  /// @param[in] key NVS Key of the string to read
+  /// @param[out] value string to read
+  /// @param[in] default_value Default value to return if key is not found
+  /// @param[out] ec Saves a std::error_code representing success or failure
+  void get(const char *key, std::string &value, const std::string &default_value,
+           std::error_code &ec) {
+    if (!check_handle_initialized(ec))
+      return;
+
+    if (!check_key(key, ec))
+      return;
+
+    esp_err_t err;
+    std::size_t len = 0;
+    err = handle_->get_item_size(nvs::ItemType::SZ, key, len);
+    if (err != ESP_OK || len == 0) {
+      logger_.warn("The value is not initialized yet! key '{}' set to: {}", key, default_value);
+      set(key, default_value, ec);
+      if (ec)
+        return;
+      value = default_value;
+    } else {
+      value.resize(len);
+      err = handle_->get_string(key, value.data(), len);
+      if (err != ESP_OK) {
+        ec = make_error_code(NvsErrc::Read_NVS_Failed);
+        logger_.error("Error {} reading from NVS!", esp_err_to_name(err));
+        return;
+      }
+    }
+  }
+
   /// @brief Save a variable in the NVS
   /// @param[in] key NVS Key of the variable to read
   /// @param[in] value Variable to read
   /// @param[out] ec Saves a std::error_code representing success or failure
   /// @details Saves the key/variable pair without committing the NVS.
   template <typename T> void set(const char *key, T value, std::error_code &ec) {
-    check_key_length(key, ec);
-    if (ec)
-      return;
-
     if (!check_handle_initialized(ec))
       return;
 
+    if (!check_key(key, ec))
+      return;
+
     esp_err_t err;
-    err = handle->set_item(key, value);
+    err = handle_->set_item(key, value);
     if (err != ESP_OK) {
       ec = make_error_code(NvsErrc::Write_NVS_Failed);
       logger_.error("Error {} writing to NVS!", esp_err_to_name(err));
@@ -196,15 +306,14 @@ public:
   /// @param[in] value string to set
   /// @param[out] ec Saves a std::error_code representing success or failure
   void set(const char *key, const std::string &value, std::error_code &ec) {
-    check_key_length(key, ec);
-    if (ec)
-      return;
-
     if (!check_handle_initialized(ec))
       return;
 
+    if (!check_key(key, ec))
+      return;
+
     esp_err_t err;
-    err = handle->set_string(key, value.data());
+    err = handle_->set_string(key, value.data());
     if (err != ESP_OK) {
       ec = make_error_code(NvsErrc::Write_NVS_Failed);
       logger_.error("Error {} writing to NVS!", esp_err_to_name(err));
@@ -217,7 +326,10 @@ public:
   /// @param[out] ec Saves a std::error_code representing success or failure
   /// @details Commits changes to the NVS
   void commit(std::error_code &ec) {
-    esp_err_t err = handle->commit();
+    if (!check_handle_initialized(ec))
+      return;
+
+    esp_err_t err = handle_->commit();
     if (err != ESP_OK) {
       logger_.error("Error {} committing to NVS!", esp_err_to_name(err));
       ec = make_error_code(NvsErrc::Commit_NVS_Failed);
@@ -226,18 +338,19 @@ public:
   }
 
 protected:
-  std::unique_ptr<nvs::NVSHandle> handle;
+  std::unique_ptr<nvs::NVSHandle> handle_;
 
-  void check_key_length(const char *key, std::error_code &ec) {
+  bool check_key(const char *key, std::error_code &ec) {
     if (strlen(key) > 15) {
       logger_.error("Key too long, must be <= 15 characters: {}", key);
       ec = make_error_code(NvsErrc::Key_Length_Too_Long);
-      return;
+      return false;
     }
+    return true;
   }
 
   bool check_handle_initialized(std::error_code &ec) {
-    if (!handle) {
+    if (!handle_) {
       ec = make_error_code(NvsErrc::Handle_Uninitialized);
       logger_.error("NVS Handle not initialized!");
       return false;
@@ -250,5 +363,5 @@ protected:
    */
   std::error_code make_error_code(NvsErrc e) { return {static_cast<int>(e), theNvsErrCategory}; }
 
-}; // Class NVSHandle
+}; // Class NvsHandle
 } // namespace espp
